Addition of roles table #6

+++ App/Controllers/RunnerController.php
@@ -6,9 +6,9 @@ use App\Core\AControllerBase;
 use App\Core\HTTPException;
 use App\Core\Responses\Response;
 use App\Helpers\FormChecker;
+use App\Helpers\PermissionChecker;
 use App\Models\Login;
-use App\Models\PersonalDetail;
-use App\Models\Runner;
+use App\Models\UserRole;
 
 class RunnerController extends AControllerBase
 {
@@ -89,7 +89,13 @@ class RunnerController extends AControllerBase
         }
 
         $login = Login::getOne($this->app->getAuth()->getLoggedUserId());
-        //TODO remove user roles
+
+        if (!PermissionChecker::canDeleteAccount($login))
+        {
+            throw new HTTPException(403, "Forbidden");
+        }
+
+        UserRole::removeAllRolesFromUser($login);
         $login->delete();
         $this->app->getAuth()->logout();
         return $this->redirect($this->url("home.index"));
